+++ src/main/java/lottos/README.md
@@ -2,7 +2,7 @@
 
 - [x] 로또 구매금액 입력
 - [ ] 로또 발급
-  - [ ] 로또 생성
+  - [x] 로또 생성
   - [x] 로또 발급
   - [ ] 입력한 금액만큼의 구매 현황 UI 출력
 - [ ] 지난 주 당첨 번호 입력
