+++ src/test/java/lottos/LottoTest.java
@@ -1,19 +1,21 @@
 package lottos;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
+import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class LottoTest {
+class LottoTest {
 
-    @DisplayName("로또 발급 로직")
-    @ParameterizedTest
-    @CsvSource(value = {"14000:14", "10000:10"}, delimiter = ':')
-    void 로또발급_1장당_1000원(final String input, final String result) {
-        Shop shop = new Shop();
-        Lottos lottos = shop.issue(input);
-        assertEquals(lottos.size(), Integer.parseInt(result));
+    @Test
+    @DisplayName("로또 생성")
+    void 로또_랜덤_생성() {
+        Lotto lotto = new Lotto();
+
+        assertEquals(lotto.getElements().size(), 6);
+        for (Integer number : lotto.getElements()) {
+            assertTrue(number >= 1 && number <= 45);
+        }
     }
 }
